ai-agents/requirements.txt
# Odin's Almanac AI Agents - Python Requirements
# ===============================================
# Dependencies for the AI-powered restaurant intelligence system

# Microsoft Agent Framework (Core AI agent functionality)
agent-framework-azure-ai==1.0.0b251016  # Using latest beta version

# OpenAI SDK for GitHub Models integration
openai>=1.12.0

# Data processing and utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
numpy>=1.24.0
pandas>=2.0.0

# Azure SDK components (for future Azure AI Foundry integration)
azure-identity>=1.15.0
azure-ai-inference>=1.0.0b1

# Logging and debugging
loguru>=0.7.0

# Optional: For advanced data analysis
scikit-learn>=1.3.0
matplotlib>=3.7.0
plotly>=5.17.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.8.0